Rely on compiler to zero out structs.
Before, we were relying on having contiguous structs and remembering
to manually zero-out fields. Alleviate all concerns by using C99
designator initializers. These are technically not slated to be added
to the C++ standard until C++20. However, because they are in C99, and
they are awesome, they are commonly used in Android.

Bug: 130161842
Test: manual
Change-Id: I96e4bea47f99e00ef33cfce74621e5869b1ff3d7