Hide TextOutput.
Avoid additional users.

Removing these completely is infeasible/not worth it for now for
libbinder, and it's important to keep libbinder/libhwbinder at least
somewhat close to each other. If saving this memory becomes important
here or if we remove this from libbinder, this ensures that no
additional work will be needed.

Bug: 148676615
Test: N/A
Change-Id: I8fd2c2c21bfc5ab47c362cdad93de9a432591a50